// lib/api/generated/model/userDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Immoplus
 * Immoplus
 * OpenAPI spec version: 1.0
 */
import type { RoleDto } from './roleDto';
import type { UserDtoAdditionalData } from './userDtoAdditionalData';
import type { UserStatus } from './userStatus';

export interface UserDto {
  id: string;
  firstName: string;
  lastName: string;
  email?: string;
  password: string;
  birthDate?: string;
  googleId: string;
  facebookId: string;
  role: RoleDto;
  language: string;
  avatar: string;
  phoneNumber: string;
  otp: string;
  otpExpiration: string;
  country: string;
  state: string;
  city: string;
  commune: string;
  address: string;
  address2: string;
  currency: string;
  additionalData: UserDtoAdditionalData;
  identityVerified: boolean;
  emailVerified: boolean;
  phoneNumberVerified: boolean;
  compteProValide: boolean;
  authLoginAttempts: number;
  status: UserStatus;
  createdAt: string;
  createdBy: string;
  updatedAt: string;
  updatedBy: string;
  deletedAt: string;
  deletedBy: string;
}